+++ CHANGELOG.md
@@ -8,6 +8,8 @@ description: |-
 # VKCS Provider's changelog
 #### v0.5.1 (unreleased)
 - Remove default sdn value from firewall and networking resources and data sources
+- Support updating Kubernetes cluster labels without recreating it
+- Add "all_labels" attribute to Kubernetes cluster
 - Wait for volume to finish detaching when deleting resource vkcs_compute_volume_attach
 
 #### v0.5.0